there was a bug in the rest controller. PUT method setter to edit books was expecting an integer as argument

+++ librarymanagementsystem/src/main/java/io/bootcamp/vimtages/model/Book.java
@@ -40,7 +40,7 @@ public class Book extends AbstractModel {
         return ISBN;
     }
 
-    public void setISBN(Integer ISBN) {
+    public void setISBN(long ISBN) {
         this.ISBN = ISBN;
     }
